chore: update references to preview commands
