+++ packages/plugin-core/src/services/NoteSyncService.ts
@@ -16,7 +16,7 @@ import _ from "lodash";
 import path from "path";
 import visit from "unist-util-visit";
 import * as vscode from "vscode";
-import { ShowPreviewV2Command } from "../commands/ShowPreviewV2";
+import { ShowPreviewCommand } from "../commands/ShowPreview";
 import { Logger } from "../logger";
 import { VSCodeUtils } from "../utils";
 import { getExtension, getDWorkspace } from "../workspace";
@@ -183,7 +183,7 @@ export class NoteSyncService {
 
     this.L.debug({ ctx, fname, msg: "exit" });
     const noteClean = await engine.updateNote(note);
-    ShowPreviewV2Command.refresh(noteClean);
+    ShowPreviewCommand.refresh(noteClean);
     return noteClean;
   }
 }
